Add new filter options: Pass Types: •	First pass only, •	Second pass only, •	third pass and beyond, •	**removed not first pass
Added Column for counting the pass number (aka: Pass count)(goes with passs type)
Added column for counting fixations in the pass for each region(fixation type)

Fixation Type:
•	First in region
•	Single fixation
•	Second of multiple
•	All subsequent fixation
•	Last fixation

Also Fixed next_region! The next region is only classified as the region that is next entered after the current region.
If there is a refixation in the same region, it is not classified as the 'next' region anymore.
The same logic applies to previous region.

fixating a region and then regressing back to it at a later point is now classified as a 'first fixation' in the region.
For example, if the subject fixates region 1, then fixates region 2, then fixates region 1 again, that is a first fixation on the second pass.
Or in other words, Second pass first fixation.